Define a topic for nav-infrastructure for us to "read"
In order to list topics, we need read access to at least one topic.
In nav-dev and nav-prod, this is provided by the canary topics, which all applications can read.
In nav-infrastructure, we need to create a dummy topic we can get read access too, so that we can list topics.